remove-cacheitem only if exists bugfix

+++ source/providers/onepassword/provider-onepassword.ps1
@@ -126,7 +126,9 @@ function global:Get-Vault {
         }
         else {
             if ($opVaultItemsCacheEnabled) {
-                Remove-CacheItem -Name $opVaultItemsCacheName -Key $Vault
+                if (Get-CacheItem -Name $opVaultItemsCacheName -Key $Vault -MaxAge $opVaultsCacheItemsMaxAge) {
+                    Remove-CacheItem -Name $opVaultItemsCacheName -Key $Vault
+                }
             }
         }
     }
